+++ sysbox-runc-lite/libcontainer/rootfs_linux_test.go
@@ -4,6 +4,7 @@ import (
 	"testing"
 
 	"github.com/opencontainers/runc/libcontainer/configs"
+	"github.com/opencontainers/runc/libcontainer/userns"
 )
 
 func TestCheckMountDestOnProc(t *testing.T) {
@@ -56,8 +57,8 @@ func TestNeedsSetupDev(t *testing.T) {
 			},
 		},
 	}
-	if needsSetupDev(config) {
-		t.Fatal("expected needsSetupDev to be false, got true")
+	if got, want := needsSetupDev(config), userns.RunningInUserNS(); got != want {
+		t.Fatalf("expected needsSetupDev=%v, got %v", want, got)
 	}
 }
 
@@ -71,8 +72,8 @@ func TestNeedsSetupDevStrangeSource(t *testing.T) {
 			},
 		},
 	}
-	if needsSetupDev(config) {
-		t.Fatal("expected needsSetupDev to be false, got true")
+	if got, want := needsSetupDev(config), userns.RunningInUserNS(); got != want {
+		t.Fatalf("expected needsSetupDev=%v, got %v", want, got)
 	}
 }